requirements.txt
Flask>=2.2.2
gunicorn>=20.1.0
psycopg2-binary>=2.9.9
asgiref>=3.6.0
environs>=9.5.0
asyncpg>=0.27.0
validators~=0.20.0
aiohttp~=3.8.1
requests~=2.31.0
aiofiles~=23.1.0
SQLAlchemy>=1.4.36
pydantic>=1.8.2
Werkzeug>=2.2.2
redis~=5.0.1
aiogram~=3.1.1
Pillow~=10.1.0

